Make the check for recursive method definitions less syntactic.
This is achieved by encoding information in the dictionary type
about accesses to class methods. The meaning of this information
depends on where the dictionary type occurs:

- If the type occurs to the left of a class arrow (i.e. in a
  function type), then the dictionary type encodes which class
  methods the function body must be granted access to.
  (Note that the dictionary type may over-approximate, i.e. it
  may -- and generally will -- require access to more methods
  than are actually used in the function body.)

- If the type is the type of an expression -- i.e. it appears in
  a position (in a function type) where no arrows follow, then
  the dictionary type encodes which class methods the expression
  allows access to.

The check for recursive method definitions is less recursive
because it now relies on types, specifically on the additional
information about class method accesses that is no encoded in
dictionary types.

NOTE: A cleaner way of implementing a type-based check for
recursive method definitions might be to annotate class arrows
with information about (required/granted) method accesses.